Reusing Auth setup
- Move BaseAPITest to base_api_test.py
- Add BaseAuthAPITest to base_api_test.py to handle tests that require an auth server.
