+++ tests/unit/backend/chalice/api_server/test_v1_collection_upload_link.py
@@ -1,39 +1,15 @@
 import json
-import sys
-import os
 import unittest
 from mock import patch
 from furl import furl
 
 from backend.corpora.common.utils.math_utils import GB
-from tests.unit.backend.chalice.api_server import BaseAPITest
-from tests.unit.backend.chalice.api_server.mock_auth import MockOauthServer, get_auth_token
+from tests.unit.backend.chalice.api_server.base_api_test import BaseAuthAPITest
+from tests.unit.backend.chalice.api_server.mock_auth import get_auth_token
 from tests.unit.backend.corpora.fixtures.environment_setup import EnvironmentSetup, fixture_file_path
 
 
-class TestCollectionUploadLink(BaseAPITest, unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        BaseAPITest.setUpClass()
-        cls.mock_oauth_server = MockOauthServer()
-        cls.mock_oauth_server.start()
-        assert cls.mock_oauth_server.server_okay
-
-        cls.old_path = sys.path.copy()
-        sys.path.insert(0, os.path.join(cls.corpora_api_dir, "chalicelib"))  # noqa
-        from corpora.common.corpora_config import CorporaAuthConfig
-
-        # Use the CorporaAuthConfig used by the chalice app
-        cls.auth_config = CorporaAuthConfig()
-        cls.auth_config._config["api_base_url"] = f"http://localhost:{cls.mock_oauth_server.port}"
-        cls.auth_config._config["callback_base_url"] = "http://localhost:5000"
-        cls.auth_config.update_defaults()
-
-    @classmethod
-    def tearDownClass(cls):
-        cls.mock_oauth_server.terminate()
-        sys.path = cls.old_path
-
+class TestCollectionUploadLink(BaseAuthAPITest, unittest.TestCase):
     def setUp(self):
         self.good_link = "https://www.dropbox.com/s/ow84zm4h0wkl409/test.h5ad?dl=0"
         self.dummy_link = "https://www.dropbox.com/s/12345678901234/test.h5ad?dl=0"
