+++ tests/unit/backend/chalice/api_server/test_v1_dataset.py
@@ -2,37 +2,14 @@ import json
 import os
 import unittest
 
-import sys
 from furl import furl
 
-from tests.unit.backend.chalice.api_server import BaseAPITest
+from tests.unit.backend.chalice.api_server.base_api_test import BaseAuthAPITest
 from tests.unit.backend.corpora import CorporaTestCaseUsingMockAWS
-from tests.unit.backend.chalice.api_server.mock_auth import MockOauthServer, get_auth_token
+from tests.unit.backend.chalice.api_server.mock_auth import get_auth_token
 
 
-class TestDataset(BaseAPITest, CorporaTestCaseUsingMockAWS):
-    @classmethod
-    def setUpClass(cls):
-        BaseAPITest.setUpClass()
-        cls.mock_oauth_server = MockOauthServer()
-        cls.mock_oauth_server.start()
-        assert cls.mock_oauth_server.server_okay
-
-        cls.old_path = sys.path.copy()
-        sys.path.insert(0, os.path.join(cls.corpora_api_dir, "chalicelib"))  # noqa
-        from corpora.common.corpora_config import CorporaAuthConfig
-
-        # Use the CorporaAuthConfig used by the chalice app
-        cls.auth_config = CorporaAuthConfig()
-        cls.auth_config._config["api_base_url"] = f"http://localhost:{cls.mock_oauth_server.port}"
-        cls.auth_config._config["callback_base_url"] = "http://localhost:5000"
-        cls.auth_config.update_defaults()
-
-    @classmethod
-    def tearDownClass(cls):
-        cls.mock_oauth_server.terminate()
-        sys.path = cls.old_path
-
+class TestDataset(BaseAuthAPITest, CorporaTestCaseUsingMockAWS):
     def test__post_dataset_asset__OK(self):
         bucket = self.CORPORA_TEST_CONFIG["bucket_name"]
         s3_file_name = "test_s3_uri.h5ad"
